Started rewrite using cogs, to be finished (added help commands and moderation commands)
