+++ main.py
@@ -30,154 +30,12 @@ async def on_ready():
         await bot.change_presence(activity=discord.Game(name="😴 sleeping..."))
         await asyncio.sleep(15)
 
-# Help command
+        # Load cogs
+        for filename in os.listdir('./cogs'):
+            if filename.endswith('.py'):
+            bot.load_extension(f'cogs.{filename[:-3]}')
 
-@bot.command()
-async def help(ctx):
-    embed = discord.Embed(
-        title="Comandos de Buggy",
-        description="",
-        timestamp=datetime.datetime.utcnow(),
-        color=discord.Color.blue())
-    embed.add_field(name="► Menu de ayuda", value="Tenemos 2 categorias y mas de 100 comandos creados.\n\n Lista de comandos: !help <categoria>\n Comandos en detalle: !help <comando>ﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠ")
-    embed.add_field(name="► Categorias",
-                    value='''
 
-    `c!mod` = Comandos de moderador                
-    `c!fun` = Comandos para reirse
-    `c!romance` = Comandos de romance
-
-    ''',
-                    inline="False")
-    embed.add_field(name="► Enlaces",value="[Pagina Web](http://google.com) | [Privacidad](http://youtube.com) | [Comandos]( http://wikipedia.com)",  inline="False")
-
-    await ctx.send(embed=embed)
-
-# Help Mod
-
-@bot.command()
-async def mod(message):
-    embed = discord.Embed(
-        title="Comandos de Buggy",
-        description="",
-        color=discord.Color.blue())
-    embed.add_field(name="Moderador", value="Ayuda sobre los comandos de moderadorﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠ", inline="False")
-    embed.add_field(name="**Comandos**",value='''
-    kick
-    ban
-    mute
-    ''',  inline="True")
-    embed.add_field(name="**Comandos**",value='''
-    clear
-    nuke
-    none
-    ''', inline="True")
-    embed.add_field(name="**Comandos**",value='''
-    none
-    none
-    none
-    ''')
-    await message.channel.send(embed=embed)
-
-# Help Fun
-
-@bot.command()
-async def fun(message):
-    embed = discord.Embed(
-        title="Comandos de Buggy",
-        description="",
-        color=discord.Color.blue())
-    embed.add_field(name="Diversion", value="Ayuda sobre los comandos de diversionﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠ", inline="False")
-    embed.add_field(name="**Comandos**",value='''
-    banana
-    ship
-    lucky
-    ''',  inline="True")
-    embed.add_field(name="**Comandos**",value='''
-    roll
-    8ball
-    none
-    ''', inline="True")
-    embed.add_field(name="**Comandos**",value='''
-    none
-    none
-    none
-    ''')
-    await message.channel.send(embed=embed)
-
-# Help Romance
-
-@bot.command()
-async def romance(message):
-    embed = discord.Embed(
-        title="Comandos de Buggy",
-        description="",
-        color=discord.Color.blue())
-    embed.add_field(name="Romance", value="Ayuda sobre los comandos de Romanceﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠﾠ", inline="False")
-    embed.add_field(name="**Comandos**",value='''
-    hug
-    kiss
-    none
-    ''',  inline="True")
-    embed.add_field(name="**Comandos**",value='''
-    none
-    none
-    none
-    ''', inline="True")
-    embed.add_field(name="**Comandos**",value='''
-    none
-    none
-    none
-    ''')
-    await message.channel.send(embed=embed)
-
-# clear command
-
-@bot.command()
-@commands.has_permissions(manage_messages=True)
-async def clear(ctx, amount=0):
-    await ctx.channel.purge(limit=amount + 1)
-
-# nuke command
-
-@bot.command()
-@commands.has_permissions(administrator=True)
-async def nuke(ctx, amount=12000):
-    await ctx.channel.purge(limit=amount + 12000)
-    embed = discord.Embed(
-        title="Buggy Bot",
-        description="Buggy, exploto una bomba nuclear por aciddente",
-        timestamp=datetime.datetime.utcnow(),
-        color=discord.Color.purple())
-    embed.add_field(name="Se han eliminado toda la conversacion del canal",
-                    value="Clown dice que lo siente :c",
-                    inline="False")
-    embed.set_thumbnail(
-        url=
-        "https://t3.ftcdn.net/jpg/03/33/09/88/360_F_333098851_oaiN8o7glILcg7nsT1vhSPewI1wSjzyN.jpg"
-    )
-    embed.set_image(
-        url=
-        "https://images-wixmp-ed30a86b8c4ca887773594c2.wixmp.com/f/41606d9e-1486-4bb8-a9b4-eba3eccb1b35/ddmmjz8-10a4c85f-41d7-4d16-b011-e329762165a7.gif?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1cm46YXBwOjdlMGQxODg5ODIyNjQzNzNhNWYwZDQxNWVhMGQyNmUwIiwiaXNzIjoidXJuOmFwcDo3ZTBkMTg4OTgyMjY0MzczYTVmMGQ0MTVlYTBkMjZlMCIsIm9iaiI6W1t7InBhdGgiOiJcL2ZcLzQxNjA2ZDllLTE0ODYtNGJiOC1hOWI0LWViYTNlY2NiMWIzNVwvZGRtbWp6OC0xMGE0Yzg1Zi00MWQ3LTRkMTYtYjAxMS1lMzI5NzYyMTY1YTcuZ2lmIn1dXSwiYXVkIjpbInVybjpzZXJ2aWNlOmZpbGUuZG93bmxvYWQiXX0.Gj7uC00OS8Ow8mAvS0SHWz0ecu9E_EQilFjEqK-FKrk"
-    )
-    await ctx.send(embed=embed)
-
-# ban command
-
-@bot.command(name="ban", help="command to ban user")
-@commands.has_permissions(ban_members=True)
-async def _ban(ctx, member: discord.Member, *, reason=None):
-    """ command to ban user. Check !help ban """
-    try:
-        await member.ban(reason=reason)
-        await ctx.message.delete()
-        await ctx.channel.send(
-            f'{member.name} Ha sido baneado de este servidor por la '
-            f'Razon: {reason}')
-    except Exception:
-        await ctx.channel.send(
-            f"No tienes el permiso suficiente para banear"
-        )
 
 # kick command
